Migrate charts files
