+++ docs/data/charts/axis/FormatterDemo.tsx
