+++ docs/data/charts/axis/ReverseExample.tsx
