+++ docs/data/charts/bars/BarGap.tsx
