+++ docs/data/charts/pie/PieShape.tsx
@@ -3,7 +3,7 @@ import ChartsUsageDemo from 'docsx/src/modules/components/ChartsUsageDemo';
 import { PieChart } from '@mui/x-charts/PieChart';
 import { desktopOS, valueFormatter } from './webUsageStats';
 
-export default function PieShapeNoSnap() {
+export default function PieShape() {
   return (
     <ChartsUsageDemo
       componentName="Pie shape"
